Fix small bugs in RC model (#191)
* MOD correct a_gold and a_pred positions

a_gold and a_pred is mistakenly reserved accroding to L69

* ADD datatype for tensor

ADD datatype for tensor to avoid it is mistakenly assigned as int or bool

* ADD is_training to make predictor work

ADD is_training to make predictor work when self.skip_impossible_questions = True

* MOD black the format

+++ allennlp_models/rc/metrics/squad_em_and_f1.py
@@ -39,7 +39,7 @@ class SquadEmAndF1(Metric):
                 device = torch.device("cpu")
             # Converting bool to int here, since we want to count the number of exact matches.
             _exact_match = torch.tensor(exact_match, dtype=torch.int).to(device)
-            _f1_score = torch.tensor(f1_score).to(device)
+            _f1_score = torch.tensor(f1_score, dtype=torch.double).to(device)
             _count = torch.tensor(count).to(device)
             dist.all_reduce(_exact_match, op=dist.ReduceOp.SUM)
             dist.all_reduce(_f1_score, op=dist.ReduceOp.SUM)
